Deudas facturación

+++ ISO2023/src/main/java/edu/uclm/esi/iso/ISO2023/services/ClienteService.java
@@ -9,9 +9,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import edu.uclm.esi.iso.ISO2023.dao.ClienteDAO;
+import edu.uclm.esi.iso.ISO2023.dao.ReservaDAO;
 import edu.uclm.esi.iso.ISO2023.dao.TokenDAO;
-import edu.uclm.esi.iso.ISO2023.entities.Token;
 import edu.uclm.esi.iso.ISO2023.entities.Cliente;
+import edu.uclm.esi.iso.ISO2023.entities.Reserva;
 import edu.uclm.esi.iso.ISO2023.exceptions.*;
 
 @Service
@@ -24,6 +25,8 @@ public class ClienteService {
 	private TokenDAO tokenDAO;
 	@Autowired
 	private SeguridadService comprobarSeguridad;
+	@Autowired
+	private ReservaDAO reservaDAO;
 	
 	private static final String ADMIN = "admin";
 
@@ -84,7 +87,12 @@ public class ClienteService {
 	}
 
 	public void darDeBaja(String email, String password) {
-		if (userService.checkUser(email, password).equals("cliente")) {
+		Optional<Cliente> deudas = this.clienteDAO.findByEmail("Deudas");
+		if (userService.checkUser(email, password).equals("cliente")&&deudas.isPresent()) {
+			for(Reserva r:this.reservaDAO.findByUsuarioEmail(email)) {
+				r.setUsuario(deudas.get());
+				this.reservaDAO.save(r);
+			}
 			this.tokenDAO.deleteAllByUserEmail(email);
 			clienteDAO.deleteById(email);
 		} else {
